break out api tests by group
