+++ api_test.go
@@ -3,15 +3,12 @@
 package hydrocarbon_test
 
 import (
-	"bytes"
 	"context"
 	"net/http"
 	"net/http/httptest"
-	"net/http/httputil"
 	"testing"
 
 	"github.com/fortytw2/hydrocarbon"
-
 	"github.com/fortytw2/hydrocarbon/pg"
 )
 
@@ -23,9 +20,10 @@ func TestAPI(t *testing.T) {
 }
 
 type apiCase struct {
-	name string
-	req  func(t *testing.T, basePath string) *http.Request
-	resp func(t *testing.T, em *hydrocarbon.MockMailer, w *httptest.ResponseRecorder)
+	name   string
+	req    func(t *testing.T, basePath string) *http.Request
+	resp   func(t *testing.T, em *hydrocarbon.MockMailer, w *httptest.ResponseRecorder)
+	authed bool
 }
 
 func runCases(t *testing.T, db *pg.DB, cases []apiCase) {
@@ -42,12 +40,14 @@ func runCases(t *testing.T, db *pg.DB, cases []apiCase) {
 			"http://localhost:3000",
 		)
 
-		ak := getAuthKey(t, db, ks)
 		w := httptest.NewRecorder()
 
 		t.Run(tt.name, func(t *testing.T) {
 			req := tt.req(t, "http://localhost:3000")
-			req.Header.Set("X-Hydrocarbon-Key", ak)
+			if tt.authed {
+				ak := getAuthKey(t, db, ks)
+				req.Header.Set("X-Hydrocarbon-Key", ak)
+			}
 
 			h.ServeHTTP(w, req)
 			tt.resp(t, mm, w)
@@ -71,25 +71,4 @@ func getAuthKey(t *testing.T, db *pg.DB, ks *hydrocarbon.KeySigner) string {
 		t.Fatal(err)
 	}
 	return signed
-}
-
-func feedApiTests(db *pg.DB) func(t *testing.T) {
-	var cases = []apiCase{
-		{
-			name: "valid-create-feed",
-			req: func(t *testing.T, baseDomain string) *http.Request {
-				return httptest.NewRequest(http.MethodPost,
-					baseDomain+"/v1/feed/create",
-					bytes.NewBufferString(`{"name": "hc", "plugin": "ycombinators", "url": "https://ycombinator.com"}`))
-			},
-			resp: func(t *testing.T, mm *hydrocarbon.MockMailer, w *httptest.ResponseRecorder) {
-				buf, _ := httputil.DumpResponse(w.Result(), true)
-				println(string(buf))
-			},
-		},
-	}
-
-	return func(t *testing.T) {
-		runCases(t, db, cases)
-	}
 }
